pyproject.toml
[project]
name = "cli-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "gradio>=6.7.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
